Fix architecture violation move formatter down to model, only option
