+++ src/main/java/org/jabref/model/cleanup/NormalizeNewlinesFormatter.java
@@ -1,10 +1,9 @@
-package org.jabref.logic.formatter.bibtexfields;
+package org.jabref.model.cleanup;
 
 import java.util.Objects;
 
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.util.OS;
-import org.jabref.model.cleanup.Formatter;
 
 /**
  * Trim all whitespace characters (defined in java) in the string.
